add doxygen for exported functions in jlapi.c
This adds some brief doxygen-formatted docstrings to jlapi.c.
The objective of this PR is to set doxygen docstrings as the
documentation method for Julia's C code, and allow for
further development of reference material for the use
of Julia through the C API.

These docstrings are immediately useful for consumption by
C/C++ development extensions such as tooltips.
In the future these can be useful to generate a cohesive API reference,
though this is a seperate concern involving the build system.

Co-authored-by: Jameson Nash <vtjnash@gmail.com>