Specialize eps(::IEEEFloat) (#52087)
`eps(::IEEEFloat)` can be made notably faster. I doubt this function
appears in performance-critical places, but better is better and this
new version is branch-free.

```julia-repl
julia> using BenchmarkTools

julia> x = randn(2^10);

julia> @btime map(eps,$x); # v1.9.1
  2.067 μs (1 allocation: 8.12 KiB)

julia> @btime map(eps,$x); # this PR
  613.095 ns (1 allocation: 8.12 KiB)
```

A couple probably-negligible caveats that arise from the use of floating
point arithmetic in this implementation:
- The previous version always returned Julia's canonical `NaN` when the
result would be NaN. In this PR, the NaN payload depends on the payload
behavior of the hardware and the input.
- The result is vulnerable to flush-to-zero floating point environments,
in which case subnormal results will be flushed to `zero(T)`. The
usability of a subnormal result amidst flushing mode is questionable, as
is the well-posed-ness of `eps`'s semantics in such a situation. *Note
that the existing implementation has this same vulnerability (due to
`ldexp`), though only when the input to `eps` is normal and the result
would be subnormal (only with an input of zero or a subnormal is a
correct subnormal output produced, assuming the flushing occurs during
arithmetic but not storage).*

Canonicalizing `NaN` would require a simple `ifelse` switch on the
result, though I can't imagine we care. Guaranteeing a proper subnormal
result in flush-to-zero mode may be more challenging (and may warrant a
branch) and is only handled slightly more correctly by the preexisting
implementation. I imagine that flushing breaks a lot more functions than
just this one, anyway.

---------

Co-authored-by: Oscar Smith <oscardssmith@gmail.com>