sroa: Handle looking through chains of KeyValue instances (#52369)
Addresses an outstanding todo from the KeyValue PR and allows (once all
the PRs are merged), optimization when multiple ScopedValues are used
`with(a=>1, b=>2)`, etc. To facilitate this, in addition to the sroa
adjustment, the ScopedValue code is adjusted to unroll the
PersistentDict creation so that the optimizer can see the full chain (we
do not support loops in the optimizer).