Fix minimum/maximum over dimensions with missing values (#35323)
`v0 != v0` returns `missing` for missing values. Use the
largest/smallest non-missing value to initialize the array. This is an
inefficient approach. Faster alternatives would be to avoid using an
initial value at all, and instead keep track of whether a value has been
set in a separate mask; or to use `typemax`/`typemin` for types that
support them.

Fixes #35308.