better support threads in precompile and jl_task_wait_empty (#52445)
Add a `nrunning` counter which identifies (when zero) when there is
nothing running anymore. Allowing us to gate all tasks on all threads on
reaching a quiescent state, not just thread 0. This should let us better
support running precompile with threads (since we will be ensured that
all of them are asleep in a consistent state before serialization tries
to inspect the process state). We could additionally stop them
afterwards to make sure there is no way for them to begin running, even
if we forgot about some other event source, but that seems unnecessary
paranoia for now.

Note it is quite hard to encounter currently, as most places where
precompile happens currently try to force the number of threads to 1.
But this should become more relevant in the future as more threads are
supported in more places. This also may help generally with being able
to ensure the IO loop is running on at least one thread (as that is
currently lacking in this PR and on master). And also help with being
able to decide on a more advanced tree-wakeup strategy, as we start to
track how many threads are in various states of running and sleeping,
relative to the amount of work they find.

Fixes #52435