Fix interaction between try/catch elision and scope handling
Currently, our try/catch elision code does not look at whether
the EnterNode has a scope, and just deletes it, if it can prove
the contents nothrow. This is obviously problematic, so this
fixes that case up to instead set the catch dest to 0 in that
case. The idea is that a later pass could use the notasklocalstate
effect to delete this after appropriate verification, but this
is not implemented. Additionally, this currently bails on folding
current_scope inside such try/catch regions - for such regions
we cannot use the CFG to determine the extent of the try/catch
region. A proper treatment of this would probably be to just
treat this as a single value mutable - but again this is not
implemented. The primary purpose of this patch is to ensure
correctness.