add atomics builtins for globals, memory, and setonce
Adds an AtomicMemory{T} type (GenericMemory{:atomic,T,Core.CPU}) for
doing element-wise atomic operations on an array (with a lock
per-element if required).

Adds the full complement of atomic operations for global variables.

Also add a *setonce set of functions for atomically setting a value, but
only if it is unset. The complementary operation (Atomic `_unsetindex!`
for AtomicMemory) is not yet provided.