sroa: Better current_scope folding
Third time's the charm hopefully. This builds on #52527 and addresses
the TODO left there by keeping track of the appropriate :scope in sroa
as we go along in order to ensure correct current_scope folding, even
when the try body is nothrow (which was left as a todo). As a result,
we also no longer result on lowering assumptions for this transformation,
which enables us to enable the lowering optimization in #52593.