Lowering: Insert QuoteNode for captured boxed value (#52596)
`Core.Box` is not self-quoting so it should be captured in a QuoteNode.
This has been benign until the improved effect system, we handle
`QuoteNode`
of a mutable value as a global access, whereas a direct reference to a
value
is treated as inaccessible memory.

Fixes #52531

---------

Co-authored-by: Gabriel Baraldi <baraldigabriel@gmail.com>
Co-authored-by: Shuhei Kadowaki <40514306+aviatesk@users.noreply.github.com>
(cherry picked from commit 1290b51a478602f4672e481ba6e5b1eed706843a)