lowering: Optimize lowering of tryfinally with trivial finally block (#52593)
This optimizes the lowering of a tryfinally block with empty finally
block to instead use the try/catch lowering, where the catch block is
given as simply `rethrow()`. This is equivalent semantically to
try/finally in this case, but the code structure is a lot simpler with
fewer basic blocks and without the auxiliary slot for tracking the
finally slot. The motivation here is to help the compiler optimize
better when using the `@with` macro, which has an empty `finally` block
(but uses the scope argument of 'tryfinally). The only problem with this
is that it violates the lowering assumptions I made in
https://github.com/JuliaLang/julia/pull/52527, so we'll probably need to
fix that first.