cfg_simplify: Add one more check to avoid merging :leave terminator (#52621)
In #52608, I made `:leave` a proper terminator (we already considered it
as such during CFG construction, we just didn't maintain that property
in the optimizer). As part of this, I adjusted one place in cfg_simplify
to avoid merging blocks that end with non-trivial terminators
(previously only EnterNode was considered). Turns out there's another
one, so fix that as well.