Make AnnotatedIOBuffer reading more generic
While `String` is the only concrete type for which `read(::IOBuffer,
::Type{<:AbstractString})` is defined, is is entirely conceivable that
some other custom string type could define a similar `read` method.

Since making reading an `AnnotatedString` from an `AnnotatedIOBuffer`
more generic is as easy as replacing the hardcoded `String` with a type
parameter, we may as well do so.