Add wrapping shift function
In Julia, bitshifting a bitinteger x by more bits than is present in x causes x
to be set to zero. Also, negative bitshifts are supported.
This might semantically be more correct (and the former also matches LLVM's
definition of bitshifts), but it does not correspond to either x86 or AArch64
behaviour. The result is that Julia's bitshifts are not optimised to a single
instruction.
In constrast, in the C language, bitshifting by more than the bitwidth is UB,
which allows the compiler to assume it never happens, and optimise the shift to
a single instruction.

This commit adds the wrapping shift functions >>% >>>% and <<%, which overflows
if the shift is too high or negative. The overflow behaviour is explicitly not
stated, but the implemented behaviour matches the native behaviour of x86 and
AArch64.

This commit requires JuliaSyntax support, which will be implemented in a parallel
PR to JuliaSyntax.