Extensions: make loading of extensions independent of what packages are in the sysimage
When triggers of extension are in the sysimage it is easy to end up with cycles in package loading.
Say we have a package A with exts BExt and CExt and say that B and C is in the sysimage.

- Upon loading A, we will immidiately start to precompile BExt (because the trigger B is "loaded" by virtue of being in the sysimage).
- BExt will load A which will cause CExt to start precompiling (again because C is in the sysimage).
- CExt will load A which will now cause BExt to start loading and we get a cycle.

This is fixed in this PR by instead of looking at what modules are loaded, we look at what modules are actually `require`d
and only use that to drive the loading of extensions.