docs: Clarify that `donotdelete` does not affect unreachable code (#52869)
In generally accepted compiler terminology, dead code is all code that
if removed does not affect the observable behavior of the program.
However, people sometimes use the phrase dead code to mean *unreachable*
code, which is a subset of dead code (being dead because it is never
semantically executed). If one assumes that definition, the docstring
for `donotdelete` may be confusing, as it may in fact be deleted from
the code if it is unreachable (existence or non-existence in the IR is
not ever semantically observable in Julia, so there could not be such an
intrinsic, but of course knowing that requires a deep understanding of
Julia semantics). Add an extra note to the docs to clarify this point.