sroa: Fix affinity of inserted undef check (#52864)
There is a bit of complexity here making sure that this instruction ends
up in the correct basic block because there's a fair number of corner
cases (e.g. whether we're deleting the first instruction of a basic
block or the last or both). Side-step all that by never doing the
delete-then-insert dance and instead doing regular instruction
replacement if necessary. Fixes #52857.