Validate more on assertion builds, and abort on errors (#52830)
This PR add a `jl_is_assertsbuild` function (cfr. `jl_is_debugbuild`)
for checking if this is an assertions build.
It's used in two places:

1. to enable additional validation, which currently often requires`-g2`
2. to make internal compiler errors fatal

Both these changes are intended to help PkgEval in detecting more issues
(invalid Julia IR) and make for more robust detection (a fatal signal
instead of having to grep logs for `Internal error`, which leads to
false positives).