lu: add allowsingular keyword argument
Currently the `check` argument conflates failed factorization (invalid
factors) and valid factorizations that have a rank-deficient factor, and
the `show` value for the factorization is misleading in the latter case.

This adds an `allowsingular` keyword argument to `lu` and the
corresponding `issuccess` methods to treat valid but rank-deficient
factors as success. It also changes `show` to show the factors as usual
in such cases, but with a "rank-deficient" note.

Fix #27657