methods: add a lock around method table changes
Ensures that only one update can occur simultaneously, and that the
acquire-release on the world-age will be sequenced after all of the
invalidations to the caches.

It should now be safe to add methods in parallel, concurrently with
running code. However, there are still no locks here to ensure that only
one module is deserialized at a time, which means that is still unsafe.

As future work, all of the method inserting and (separately) the
invalidation work could run in parallel after loading an image, since
there is a fine-grained lock on each individual part of that, as well as
the big lock so that only one module at a time can load methods.

The min/max are always written behind a lock, but they are not usually
read from behind a lock, so add relaxed atomic markers to make any
analyzers happy.