refactor the optimization flags set and utilities
This commit serves as a preliminary PR for #52991. It involves adding
`IR_FLAG_TERMINATES` to the optimization flags and carrying out various
renaming refactors. As `IR_FLAG_TERMINATES` isn't utilized in this
particular commit, it doesn't bring any changes to the compiler's
functionality. The incoming commit will make use of it and fix #52991
along with other accompanying changes.