Generalize `==` Documentation to not be Statastics-Specific
The `==` documentation made a special-case for the statistics community, specifically allowing a separate contract on `missing`. This denies extendability and only focuses the needs of only one small subset of the Julia community, which is something that is generally frowned upon.

This change to the documentation instead describes the generalization of this process, where `(::T1 == ::T2)::Union{Bool, promote_type(T2)}`, which is a version of the contract that includes the `missing` case as part of the general rule, but covers other areas of the Julia community, including symbolic computing, statically-defined boolean types (Static.jl, FillArrays.jl), and makes a clear rule beyond just handling the `missing` special case.