Introduce new syntax highlighting stdlib (#51810)
Having this as a standard library allows for syntax highlighting to be
used in the Markdown and REPL standard libraries. As a bonus, this
provides an easy and canonical way of obtaining syntax highlighted code
across the ecosystem.

This is a pretty tiny library, at just ~200 lines, but a useful one.