reorder the code in LinearAlgebra.wrap to be slightly more inferrable abstractly
By adding a type-assert at the end, we can try to ensure that eltype and
ndims info is preserved, even if inference decides not to preserve the
giant Union in detail. This avoids complete loss of precision in the
subsequent calls (generic_matmatmul! -> _generic_matmatmul!), which has
a type-assert on each input of AbstractVecOrMat.

Closes #53075