Check with is_in_mmtk_spaces instead of is_mapped_address (mmtk/mmtk-julia#123)
Objects in VM spaces are not 'mmapped' by MMTk, so `is_mapepd_address`
will return `false` for those. But they are still considered as in MMTk
spaces (the VM space).

Note that malloc'd objects are not mapped by MMTk, and not in MMTk
spaces, so both methods will return `false` for malloc'd objects.
However, this seems irrelevant in this scan object code.