Implement empty! for Channel
What is says on the box. Returning the number of elements removed is analogous to `write`, and can provide useful information.

I have not changed any changelog. I am also not certain if I should have used any `@atomic` annotations, which I saw a lot of in this file. But I would not know how.