further improve clarity of MethodError printing
Distinguish some of the cases of manually thrown MethodError by looking
for a method in the specified world. Also refactor some of the kwcall
handling to be closer to supporting `invoke`d calls (although it does
not guaranteed to have a value for `f`, which is often required later).

Fixes #36182